refactor: update splash view and introduce splash view model
